+++ .eslintrc.js
@@ -9,6 +9,7 @@ module.exports = {
     'import',
     'jest',
     'simple-import-sort',
+    'unicorn',
   ],
   env: {
     es6: true,
@@ -40,6 +41,11 @@ module.exports = {
     tsconfigRootDir: __dirname,
     warnOnUnsupportedTypeScriptVersion: false,
     EXPERIMENTAL_useSourceOfProjectReferenceRedirect: false,
+    cacheLifetime: {
+      // we pretty well never create/change tsconfig structure - so need to ever evict the cache
+      // in the rare case that we do - just need to manually restart their IDE.
+      glob: 'Infinity',
+    },
   },
   rules: {
     // make sure we're not leveraging any deprecated APIs
@@ -106,6 +112,13 @@ module.exports = {
     //
 
     curly: ['error', 'all'],
+    eqeqeq: [
+      'error',
+      'always',
+      {
+        null: 'never',
+      },
+    ],
     'no-mixed-operators': 'error',
     'no-console': 'error',
     'no-process-exit': 'error',
@@ -186,6 +199,10 @@ module.exports = {
 
     // enforce a sort order across the codebase
     'simple-import-sort/imports': 'error',
+
+    'one-var': ['error', 'never'],
+
+    'unicorn/no-typeof-undefined': 'error',
   },
   overrides: [
     // all test files