+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/intrinsic-keyword/fixture.ts
@@ -1,0 +1,6 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+type Uppercase<S extends string> = intrinsic;
+type Lowercase<S extends string> = intrinsic;
+type Capitalize<S extends string> = intrinsic;
+type Uncapitalize<S extends string> = intrinsic;