+++ packages/eslint-plugin/src/rules/prefer-includes.ts
@@ -1,7 +1,7 @@
+import type { AST as RegExpAST } from '@eslint-community/regexpp';
+import { parseRegExpLiteral } from '@eslint-community/regexpp';
 import type { TSESLint, TSESTree } from '@typescript-eslint/utils';
 import { AST_NODE_TYPES } from '@typescript-eslint/utils';
-import type { AST as RegExpAST } from 'regexpp';
-import { parseRegExpLiteral } from 'regexpp';
 import * as ts from 'typescript';
 
 import {
@@ -38,7 +38,7 @@ export default createRule({
 
     function isNumber(node: TSESTree.Node, value: number): boolean {
       const evaluated = getStaticValue(node, globalScope);
-      return evaluated !== null && evaluated.value === value;
+      return evaluated != null && evaluated.value === value;
     }
 
     function isPositiveCheck(node: TSESTree.BinaryExpression): boolean {