+++ packages/shared-fixtures/fixtures/javascript/destructuring/class-method-params-defaults-array.src.js
@@ -1,4 +1,0 @@
-class A {
-    foo([bar=3, baz=4]) {
-    }
-}