+++ packages/shared-fixtures/fixtures/typescript/basics/call-signatures-with-generics.src.ts
@@ -1,4 +1,0 @@
-type foo = {
-  <T>(a: string): string
-  new<T>(a: string): string
-}