+++ packages/shared-fixtures/fixtures/typescript/basics/class-with-override-method.src.ts
@@ -1,5 +1,0 @@
-class SpecializedComponent extends SomeComponent {
-  override show() {
-      // ...
-  }
-}