+++ packages/shared-fixtures/fixtures/typescript/basics/null-and-undefined-type-annotations.src.ts
@@ -1,2 +1,0 @@
-let x: null;
-let y: undefined;