+++ packages/shared-fixtures/fixtures/typescript/basics/type-assertion-in-arrow-function.src.ts
@@ -1,3 +1,0 @@
-const assertString = (x: any): asserts x => {
-    return
-}