+++ packages/shared-fixtures/fixtures/typescript/decorators/parameter-decorators/parameter-rest-element-decorator.src.ts
@@ -1,3 +1,0 @@
-class Foo {
-  bar(@special(true) ...foo: any) {}
-}