+++ packages/shared-fixtures/fixtures/typescript/errorRecovery/interface-index-signature-private.src.ts
@@ -1,4 +1,0 @@
-interface Foo {
-  private [baz: string]: string;
-}
-