+++ packages/shared-fixtures/fixtures/typescript/errorRecovery/interface-method-public.src.ts
@@ -1,4 +1,0 @@
-interface Foo {
-    public g(bar: string): void;
-}
-