+++ packages/typescript-estree/src/version-check.ts
@@ -24,6 +24,8 @@ const versions = [
   '4.6',
   '4.7',
   '4.8',
+  '4.9',
+  '5.0',
 ] as const;
 type Versions = typeof versions extends ArrayLike<infer U> ? U : never;